Update rsa.cu
(int)

+++ rsa.cu
@@ -45,7 +45,7 @@ __global__ void verschluessselung(int klartexte[], int geheimtexte[])
 	for (i = 0 ; i < block_length; i++)
 	{
 		//Integer hoch 103 ist zu hoch!
-		geheimtexte[i+blockIdx.x*block_length] = pow(klartexte[i+blockIdx.x*block_length],3) % 15;
+		geheimtexte[i+blockIdx.x*block_length] = (int)pow(klartexte[i+blockIdx.x*block_length],3) % 15;
 		//geheimtexte[i+blockIdx.x*block_length] = mypow(,v);
 		
 		/*
@@ -68,7 +68,7 @@ __global__ void entschluessselung(int geheimtexte[], int klartexte_pruefung[])
 	for (i = 0 ; i < block_length; i++)
 	{
 		//Integer hoch 103 ist zu hoch!
-		klartexte_pruefung[i+blockIdx.x*block_length] = pow(geheimtexte[i+blockIdx.x*block_length],e) % n;
+		klartexte_pruefung[i+blockIdx.x*block_length] = (int)pow(geheimtexte[i+blockIdx.x*block_length],e) % n;
 		
 		/*
 		multi = x  = geheimtexte[i+blockIdx.x*block_length];
